repo_path: move parse_fs_path() to jj-lib::ui_path
This function parses a "ui" path into RepoPath. file_util can also be moved back
to jj-lib if needed.
